Extract capacity queue controller to separate file
Move createCapacityQueueController and its helper functions from PhysicsControllers.ts into a dedicated CapacityQueueController.ts file to improve code organization and maintainability.

Also add validation to physicsReferenceEnvelope to reject oversized explicit sample grids exceeding the maximum of 1,000,000 points.
